리팩토링 : 주급 계산 예외 추가
