+++ src/main/resources/templates/member/createCalcForm.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<link rel="stylesheet" href="/css/createCalcForm.css">
+<head>
+    <script src="https://kit.fontawesome.com/01dfb9b9dd.js" crossorigin="anonymous"></script>
+    <meta charset="UTF-8">
+</head>
+<meta charset="UTF-8">
+
+<div style="text-align: center;">
+    <div class="title">커피홀 창동점</div>
+    <div class="mini-title"> 근무시간 입력</div>
+    <body>
+    <div class="container">
+        <form name="calcForm" action="/member/calc" method="post" th:object="${calcForm}">
+            <div>
+                <select name="memberId" id="memberId">
+                    <option> 직원 선택</option>
+                    <option th:each="member : ${members}" th:value="${member.id}"
+                            th:text="${member.name}"/>
+                </select>
+            </div>
+            <table border="1">
+                <tr>
+                    <td>1주차</td>
+                    <td>
+                        <label th:for="time1"></label>
+                        <input type="number" th:field="*{time1}"/>
+                    </td>
+                </tr>
+                <tr>
+                    <td>2주차</td>
+                    <td>
+                        <label th:for="time2"></label>
+                        <input type="number" th:field="*{time2}"/>
+                    </td>
+                </tr>
+                <tr>
+                    <td>3주차</td>
+                    <td>
+                        <label th:for="time3"></label>
+                        <input type="number" th:field="*{time3}"/>
+                    </td>
+                </tr>
+                <tr>
+                    <td>4주차</td>
+                    <td>
+                        <label th:for="time4"></label>
+                        <input type="number" th:field="*{time4}"/>
+                    </td>
+                </tr>
+                <tr>
+                    <td>5주차</td>
+                    <td>
+                        <label th:for="time5"></label>
+                        <input type="number" th:field="*{time5}"/>
+                    </td>
+                </tr>
+                <tr>
+                    <td colspan="2">
+                        <input style="width: 100%;" class="submit-button" type="submit" value="계산하기"/>
+                    </td>
+                </tr>
+                </th>
+            </table>
+        </form>
+    </div>
+    </body>
+</div>
+
+</html>
