--- src/main/resources/templates/member/createCalcMemberForm.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<style>
-    th, tr, td {
-        border: solid 1px #000;
-    }
-
-    .center-text {
-        text-align: center;
-    }
-</style>
-<meta charset="UTF-8">
-
-<div style="text-align: center;">
-    <body>
-    <h2 style="text-align: center;margin-bottom: 6%;">
-        시급과 주간 근무시간을 입력해주세요
-    </h2>
-
-    <!--<div th:each="member : ${members}" style="margin : 0 auto; width: 720px">-->
-    <form name="calcForm" action="/member/calc" method="post" th:object="${memberCalcForm}">
-        <div>
-            <label>직원 목록</label>
-            <select name="memberId" id="memberId">
-                <option value="">직원 선택</option>
-                <option th:each="member : ${members}" th:value="${member.id}"
-                        th:text="${member.name}"/>
-            </select>
-        </div>
-        <!--        <div style="margin : 0 auto; width: 720px">-->
-        <!--            <input type="hidden" th:field="*{memberId}" th:value="${member.id}"/>-->
-        <table border="1" style="display: inline-block; border-collapse: collapse; margin: 5px;">
-            <tr>
-                <td style="text-align: center;">1주차</td>
-                <td>
-                    <label th:for="time1"></label>
-                    <input type="number" th:field="*{time1}" align="right"/>
-                </td>
-            </tr>
-            <tr>
-                <td style="text-align: center;">2주차</td>
-                <td>
-                    <label th:for="time2"></label>
-                    <input type="number" th:field="*{time2}" align="right"/>
-                </td>
-            </tr>
-            <tr>
-                <td style="text-align: center;">3주차</td>
-                <td>
-                    <label th:for="time3"></label>
-                    <input type="number" th:field="*{time3}" align="right"/>
-                </td>
-            </tr>
-            <tr>
-                <td style="text-align: center;">4주차</td>
-                <td>
-                    <label th:for="time4"></label>
-                    <input type="number" th:field="*{time4}" align="right"/>
-                </td>
-            </tr>
-            <tr>
-                <td style="text-align: center;">5주차</td>
-                <td>
-                    <label th:for="time5"></label>
-                    <input type="number" th:field="*{time5}" align="right"/>
-                </td>
-            </tr>
-            <tr>
-                <td colspan="2">
-                    <input style="width: 100%;" class="submit-button" type="submit" value="계산하기"/>
-            </tr>
-            </th>
-            <!--        <p>-->
-            <!--            <label th:for="weekTime">시급</label>-->
-            <!--            <input type="text" th:field="*{hourlyWage}" align="right"/>-->
-            <!--        </p>-->
-
-        </table>
-    </form>
-    <!--</div>-->
-    </body>
-</div>
-
-</html>
